chatbot-flask-app/requirements.txt
Flask==2.2.2
SQLAlchemy==1.4.36
psycopg2-binary==2.9.3
langchain==0.0.200
openai==0.27.0
PyPDF2==1.26.0
python-dotenv==0.20.0
flask-cors==3.0.10
numpy==1.23.5
faiss-cpu==1.7.2